requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
